LPD-94965 Move helper classes to a helper package
