+++ workspaces/liferay-osbfaro-workspace/modules/osb-faro-web/src/main/java/com/liferay/osb/faro/web/internal/controller/contacts/ContactsLayoutFaroController.java
@@ -8,9 +8,9 @@ package com.liferay.osb.faro.web.internal.controller.contacts;
 import com.liferay.osb.faro.contacts.model.ContactsLayoutTemplate;
 import com.liferay.osb.faro.contacts.service.ContactsLayoutTemplateLocalService;
 import com.liferay.osb.faro.model.FaroProject;
-import com.liferay.osb.faro.web.internal.card.template.ContactsCardTemplateManagerHelper;
 import com.liferay.osb.faro.web.internal.controller.BaseFaroController;
 import com.liferay.osb.faro.web.internal.controller.FaroController;
+import com.liferay.osb.faro.web.internal.helper.ContactsCardTemplateManagerHelper;
 import com.liferay.osb.faro.web.internal.model.display.contacts.ContactsLayoutDisplay;
 import com.liferay.osb.faro.web.internal.model.display.contacts.card.template.ContactsLayoutTemplateDisplay;
 import com.liferay.osb.faro.web.internal.model.display.main.FaroEntityDisplay;
