+++ workspaces/liferay-osbfaro-workspace/modules/osb-faro-web/src/main/java/com/liferay/osb/faro/web/internal/helper/ContactsCardTemplateManagerHelper.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.osb.faro.web.internal.card.template;
+package com.liferay.osb.faro.web.internal.helper;
 
 import com.fasterxml.jackson.core.type.TypeReference;
